updates to openface extractor and zmq interface

+++ extractors/openface/extract.py
@@ -3,7 +3,6 @@ import shutil
 import subprocess
 import tempfile
 from pathlib import Path
-
 import click
 
 
@@ -21,13 +20,9 @@ def cli():
 @click.option("--openface_bin", default="FeatureExtraction")
 def extract_openface(input_file, output_dir, openface_args, openface_bin):
     with tempfile.TemporaryDirectory() as td:
-
-        # try:
         subprocess.check_call(
             [openface_bin, "-f", input_file, "-out_dir", td, openface_args, "-q"]
         )
-        # except subprocess.CalledProcessError as e:
-        #     print(e)
         input_path = Path(input_file)
         csv_file_name = os.path.join(td, input_path.with_suffix(".csv").name)
         video_file_name = os.path.join(td, input_path.with_suffix(".avi").name)
@@ -37,9 +32,14 @@ def extract_openface(input_file, output_dir, openface_args, openface_bin):
         output_csv = os.path.join(output_dir, input_path.with_suffix(".csv").name)
         output_avi = os.path.join(output_dir, input_path.with_suffix(".avi").name)
 
-        shutil.copy(video_file_name, output_avi)
-        with open(output_csv, "w") as outfile:
-            subprocess.check_call(["sed", "s/, /,/g", csv_file_name], stdout=outfile)
+        if os.path.exists(video_file_name):
+            shutil.copy(video_file_name, output_avi)
+
+        if os.path.exists(csv_file_name):
+            with open(output_csv, "w") as outfile:
+                subprocess.check_call(
+                    ["sed", "s/, /,/g", csv_file_name], stdout=outfile
+                )
 
 
 if __name__ == "__main__":
